+++ codexdispatch/__init__.py
@@ -1,0 +1,14 @@
+from .args import parse_args
+from .utils import collect_files, parse_codex_json, _invoke_codex
+from .security_audit import run_security_audit, AUDIT_TEMPLATE_PATH
+from .dispatcher import main
+
+__all__ = [
+    "parse_args",
+    "collect_files",
+    "parse_codex_json",
+    "_invoke_codex",
+    "run_security_audit",
+    "AUDIT_TEMPLATE_PATH",
+    "main",
+]
