+++ codexdispatch/dispatcher.py
@@ -1,0 +1,220 @@
+import os
+import sys
+import logging
+import uuid
+import shutil
+import subprocess
+import json
+from concurrent.futures import ThreadPoolExecutor
+from typing import Dict, List
+
+from .args import parse_args
+from .utils import collect_files, _invoke_codex
+from .security_audit import run_security_audit
+
+
+def main() -> None:
+    args = parse_args()
+
+    if args.security_audit:
+        run_security_audit(args)
+        return
+
+    template_path = args.template
+    output_dir = args.output_dir
+    workers = args.workers
+    passes = max(1, args.passes)
+    if passes > 1 and not args.tree_dirs:
+        logging.error("--passes > 1 is only supported with --tree-dirs")
+        sys.exit(1)
+    map_name = args.map_name or f"mp_map_{uuid.uuid4().hex}.json"
+
+    if args.data_dir and not args.recursive:
+        logging.warning("--recursive option is ignored when using --data-dir")
+    if args.file_list and not args.recursive:
+        logging.warning("--recursive option is ignored when using --file-list")
+
+    with open(template_path, "r") as f:
+        template = f.read()
+
+    os.makedirs(output_dir, exist_ok=True)
+
+    root_prefix: Dict[str, str] = {}
+    if args.tree_dirs:
+        for idx, d in enumerate(args.tree_dirs, 1):
+            base = os.path.basename(os.path.normpath(d))
+            root_prefix[os.path.abspath(d)] = f"{idx}_{base}"
+
+    file_list_entries: List[str] = []
+    if args.file_list:
+        if not os.path.isdir(args.work_dir):
+            logging.error("--work-dir %s does not exist", args.work_dir)
+            sys.exit(1)
+        with open(args.file_list, "r", encoding="utf-8") as f:
+            file_list_entries = [
+                p.strip()
+                for p in f
+                if p.strip() and not p.strip().startswith("#")
+            ]
+        file_list_entries = sorted(dict.fromkeys(file_list_entries))
+
+    codex_bin = args.codex_bin
+    if codex_bin:
+        codex_bin = os.path.abspath(codex_bin)
+        if not os.path.exists(codex_bin):
+            logging.error("Codex binary not found at %s", codex_bin)
+            sys.exit(1)
+    else:
+        codex_bin = shutil.which("codex")
+        if codex_bin is None:
+            candidates = [
+                f
+                for f in os.listdir(os.getcwd())
+                if os.path.isfile(f) and "codex" in f and os.access(f, os.X_OK)
+            ]
+            if len(candidates) == 1:
+                codex_bin = os.path.abspath(candidates[0])
+            elif len(candidates) > 1:
+                logging.error(
+                    "Multiple codex binaries found in current directory: %s",
+                    ", ".join(candidates),
+                )
+                sys.exit(1)
+            else:
+                logging.error("Codex binary not found in PATH or current directory")
+                sys.exit(1)
+
+    logging.basicConfig(level=logging.INFO, format="%(levelname)s: %(message)s")
+
+    def rel_and_root(path: str):
+        if args.tree_dirs:
+            root = next(
+                (
+                    d
+                    for d in args.tree_dirs
+                    if os.path.commonpath([os.path.abspath(path), os.path.abspath(d)])
+                    == os.path.abspath(d)
+                ),
+                os.path.dirname(path),
+            )
+            prefix = root_prefix.get(os.path.abspath(root), os.path.basename(root))
+            rel_path = os.path.join(prefix, os.path.relpath(path, root))
+            return rel_path, root
+        elif args.data_dir:
+            return os.path.relpath(path, args.data_dir), None
+        else:
+            return os.path.relpath(path), None
+
+    def build_prompt(data: str, prev_output: str | None) -> str:
+        if prev_output is None:
+            return f"{template}\n{data}"
+        return f"{template}\n{prev_output}\n{data}"
+
+    def run_on_file(pass_idx: int, prev_outputs: Dict[str, str], path: str) -> None:
+        try:
+            try:
+                with open(path, "r", encoding="utf-8") as f:
+                    data = f.read()
+            except UnicodeDecodeError:
+                prefix = "FileListMode:" if args.file_list else ""
+                logging.warning("%sSkipping non-text file %s", prefix, path)
+                return
+            prev_output = prev_outputs.get(path)
+            if args.file_list:
+                filename_line = os.path.abspath(path)
+                file_data = f"{filename_line}\n{data}"
+            else:
+                file_data = data
+            prompt = build_prompt(file_data, prev_output)
+
+            rel_path, root = rel_and_root(path)
+
+            file_name = f"{pass_idx}_{rel_path}-codex" if passes > 1 else f"{rel_path}-codex"
+            output_path = os.path.join(output_dir, file_name)
+            os.makedirs(os.path.dirname(output_path), exist_ok=True)
+
+            work_dir = args.work_dir if args.file_list else (
+                os.path.dirname(path) if args.tree_dirs else args.work_dir
+            )
+            cmd = [
+                codex_bin,
+                "exec",
+                "--output-last-message",
+                output_path,
+                "--dangerously-bypass-approvals-and-sandbox",
+                "--skip-git-repo-check",
+                "-C",
+                work_dir,
+            ]
+            if args.file_list:
+                logging.info("FileListMode: %s -> %s", path, output_path)
+            elif args.tree_dirs:
+                logging.info(
+                    "TreeMode: pass=%s root=%s   file=%s   work_dir=%s",
+                    pass_idx,
+                    root,
+                    path,
+                    work_dir,
+                )
+            else:
+                logging.info("Running codex on %s", path)
+
+            _invoke_codex(cmd, prompt, args.timeout, path)
+
+            prefix = "FileListMode:" if args.file_list else ""
+            logging.info("%sWrote %s", prefix, output_path)
+        except subprocess.TimeoutExpired as te:
+            logging.error("TIMEOUT after %ss on %s", te.timeout, path)
+        except subprocess.CalledProcessError as cpe:
+            stderr = (cpe.stderr or b"").decode(errors="ignore")
+            logging.error("Codex exit %s on %s\n%s", cpe.returncode, path, stderr)
+        except Exception as exc:
+            logging.error("Failed processing %s: %s", path, exc)
+
+    if args.tree_dirs:
+        files = collect_files(args.tree_dirs, recursive=args.recursive)
+    elif args.data_dir:
+        data_dir = args.data_dir
+        files = [
+            os.path.join(dp, f)
+            for dp, _, filenames in os.walk(data_dir)
+            for f in filenames
+            if os.path.isfile(os.path.join(dp, f))
+        ]
+        files = sorted(files)
+    else:
+        files = [p for p in file_list_entries if os.path.exists(p)]
+        missing = [p for p in file_list_entries if not os.path.exists(p)]
+        for m in missing:
+            logging.warning("FileListMode: missing path %s", m)
+
+    for pass_idx in range(1, passes + 1):
+        prev_cache: Dict[str, str] = {}
+        if pass_idx > 1:
+            for p in files:
+                rel_path, _ = rel_and_root(p)
+                prev_file = (
+                    f"{pass_idx - 1}_{rel_path}-codex" if passes > 1 else f"{rel_path}-codex"
+                )
+                prev_path = os.path.join(output_dir, prev_file)
+                if os.path.exists(prev_path):
+                    with open(prev_path, "r", encoding="utf-8") as pf:
+                        prev_cache[p] = pf.read()
+                else:
+                    logging.error("Missing previous output %s", prev_path)
+
+        with ThreadPoolExecutor(max_workers=workers) as executor:
+            list(executor.map(lambda p: run_on_file(pass_idx, prev_cache, p), files))
+
+        if pass_idx < passes:
+            mapping: List[dict[str, str]] = []
+            for p in files:
+                rel_path, _ = rel_and_root(p)
+                file_name = (
+                    f"{pass_idx}_{rel_path}-codex" if passes > 1 else f"{rel_path}-codex"
+                )
+                out_path = os.path.join(output_dir, file_name)
+                if os.path.exists(out_path):
+                    mapping.append({"input": os.path.abspath(p), "output": os.path.abspath(out_path)})
+            with open(os.path.join(output_dir, map_name), "w", encoding="utf-8") as mf:
+                json.dump(mapping, mf, indent=2)
